corrected code
