+++ OOP_Tasks/adding structure objects/Program.cs
@@ -12,26 +12,4 @@ namespace adding_structure_objects
             Console.WriteLine(p3.ToString());
         }
     }
-
-    public struct Point
-    {
-        int x;
-        int y;
-
-        public Point(int X, int Y)
-        {
-            x = X;
-            y = Y;
-        }
-
-        public static Point operator +(Point point1, Point point2)
-        {
-            return new Point((point1.x + point2.x), (point1.y + point2.y));
-        }
-
-        public override string ToString()
-        {
-            return $"a structure object with coordinates X = {x}, Y = {y}";
-        }
-    }
 }
